migrate Arbitrary instances and Elem to own files, cleanup
