+++ src/executor.rs
@@ -1,5 +1,6 @@
 use crate::restack::{Restack, RestackError};
-use crate::types::{Elem, ElemError, Instruction, Instructions};
+use crate::elem::{Elem, ElemError};
+use crate::types::{Instruction, Instructions};
 use thiserror::Error;
 
 #[derive(Debug, Default)]
